Change for issue #68 - Update Stranded<T> -> Stranded Add new interface Transposable<T> with <T> methods from original Stranded implementation. Update GenomicRegion to be Transposable<GenomicRegion>
