+++ src/main/java/org/monarchinitiative/svart/GenomicRegion.java
@@ -12,7 +12,7 @@ import static org.monarchinitiative.svart.GenomicComparators.*;
  * @author Jules Jacobsen <j.jacobsen@qmul.ac.uk>
  * @author Daniel Danis <daniel.danis@jax.org>
  */
-public interface GenomicRegion extends Region<GenomicRegion>, Stranded<GenomicRegion> {
+public interface GenomicRegion extends Region<GenomicRegion>, Transposable<GenomicRegion> {
 
     /**
      * @return contig where the region is located
