+++ src/main/java/org/monarchinitiative/svart/Stranded.java
@@ -3,21 +3,8 @@ package org.monarchinitiative.svart;
 /**
  * @author Jules Jacobsen <j.jacobsen@qmul.ac.uk>
  */
-public interface Stranded<T> {
+public interface Stranded {
 
     Strand strand();
 
-    T withStrand(Strand other);
-
-    default T toOppositeStrand() {
-        return withStrand(strand().opposite());
-    }
-
-    default T toPositiveStrand() {
-        return withStrand(Strand.POSITIVE);
-    }
-
-    default T toNegativeStrand() {
-        return withStrand(Strand.NEGATIVE);
-    }
 }
